refactor: Move Span -> mod error
Span is used for both the parse tree and the abstract syntax tree.
Span is used for better error reporting, so it belongs to the error
module.
